Finish basic functions
Finished getting sentence. Used a struct to store the sentence. It
stored the original sentence as a string, the words and the words
converted in a vector.

Function was added to get the individual words and remove the
punctuation.

Function to convert the words into pig-latin was added. A function
called is_vowel() was created to help convert the words.

Function was added to display the new converted words. However the
punctuation was not preserved.